requirements.txt
azure-cognitiveservices-speech
librosa
azure-ai-translation-text
azure-ai-textanalytics
line-bot-sdk
flask
google-generativeai
mysql-connector-python